Move properties back to traits
