+++ src/Fields/Properties/WithHandle.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Aerni\LivewireForms\Fields\Properties;
+
+trait WithHandle
+{
+    protected function handleProperty(): string
+    {
+        return $this->field->handle();
+    }
+}
