+++ src/Fields/Properties/WithId.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Aerni\LivewireForms\Fields\Properties;
+
+trait WithId
+{
+    protected function idProperty(): string
+    {
+        return "{$this->id}_{$this->handle()}";
+    }
+}
