+++ src/Fields/Properties/WithKey.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Aerni\LivewireForms\Fields\Properties;
+
+trait WithKey
+{
+    protected function keyProperty(): string
+    {
+        return "data.{$this->handle()}";
+    }
+}
